@@ -29,27 +29,23 @@
   "homepage": "https://github.com/foxdonut/meiosis-examples#readme",
   "dependencies": {
     "classnames": "^2.2.5",
-    "handlebars": "^4.0.5",
-    "history": "^4.3.0",
+    "handlebars": "^4.0.6",
+    "history": "^4.5.0",
     "jquery": "^3.1.1",
-    "meiosis": "^0.9.0",
-    "meiosis-mithril": "^0.8.0",
-    "meiosis-react": "^0.8.0",
-    "meiosis-snabbdom": "^0.8.0",
-    "meiosis-tracer": "^0.8.0",
-    "meiosis-vanillajs": "^0.8.0",
-    "mithril": "^0.2.5",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
+    "meiosis": "^0.10.0",
+    "meiosis-tracer": "^0.10.0",
+    "mithril": "^1.0.0-rc.6",
+    "react": "^15.4.1",
+    "react-dom": "^15.4.1",
     "snabbdom": "^0.5.4",
     "union-type": "^0.3.3"
   },
   "devDependencies": {
-    "babel-core": "^6.18.0",
-    "babel-loader": "^6.2.7",
+    "babel-core": "^6.21.0",
+    "babel-loader": "^6.2.10",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
     "cross-env": "^3.1.3",
-    "webpack": "^1.13.3"
+    "webpack": "^1.14.0"
   }
 }
